pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coverquery"
version = "0.1.0"
description = "CoverQuery CLI"
requires-python = ">=3.10"
dependencies = ["coverage>=7.0", "pyyaml>=6.0"]

[project.scripts]
coverquery = "coverquery.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
